@@ -17,19 +18,19 @@ def getProperty(String filename, String propName) {
 }
 
 android {
-    compileSdkVersion 26
-    buildToolsVersion '26.0.2'
+    compileSdkVersion 29
+    buildToolsVersion '28.0.3'
 
     defaultConfig {
        // applicationId "com.esoxjem.movieguide"
-        minSdkVersion 16
-        targetSdkVersion 26
-       // versionCode 1
-       // versionName "1.0"
+        minSdkVersion 21
+        targetSdkVersion 29
+        //versionCode 1
+        //versionName "1.0"
 
         buildConfigField "String", "TMDB_API_KEY", "\"${getProperty("local.properties", "tmdb_api_key")}\""
-        buildConfigField "String", "TMDB_BASE_URL", "\"http://api.themoviedb.org/\""
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+        buildConfigField "String", "TMDB_BASE_URL", "\"https://api.themoviedb.org/\""
+        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
     }
     buildTypes {
         release {
