@@ -3,8 +3,7 @@ package com.esoxjem.movieguide;
 /**
  * @author arun
  */
-public class Constants
-{
+public class Constants {
 
     public static final String MOVIE = "movie";
     public static final String EMPTY = "";
