@@ -6,90 +6,74 @@ import android.os.Parcelable;
 /**
  * @author arun
  */
-public class Review implements Parcelable
-{
+public class Review implements Parcelable {
     private String id;
     private String author;
     private String content;
     private String url;
 
-    public Review()
-    {
+    public Review() {
 
     }
 
-    protected Review(Parcel in)
-    {
+    protected Review(Parcel in) {
         id = in.readString();
         author = in.readString();
         content = in.readString();
         url = in.readString();
     }
 
-    public static final Creator<Review> CREATOR = new Creator<Review>()
-    {
+    public static final Creator<Review> CREATOR = new Creator<Review>() {
         @Override
-        public Review createFromParcel(Parcel in)
-        {
+        public Review createFromParcel(Parcel in) {
             return new Review(in);
         }
 
         @Override
-        public Review[] newArray(int size)
-        {
+        public Review[] newArray(int size) {
             return new Review[size];
         }
     };
 
-    public String getId()
-    {
+    public String getId() {
         return id;
     }
 
-    public void setId(String id)
-    {
+    public void setId(String id) {
         this.id = id;
     }
 
-    public String getAuthor()
-    {
+    public String getAuthor() {
         return author;
     }
 
-    public void setAuthor(String author)
-    {
+    public void setAuthor(String author) {
         this.author = author;
     }
 
-    public String getContent()
-    {
+    public String getContent() {
         return content;
     }
 
-    public void setContent(String content)
-    {
+    public void setContent(String content) {
         this.content = content;
     }
 
-    public String getUrl()
-    {
+    public String getUrl() {
         return url;
     }
 
-    public void setUrl(String url)
-    {
+    public void setUrl(String url) {
         this.url = url;
     }
 
     @Override
-    public int describeContents()
-    {
+    public int describeContents() {
         return 0;
     }
 
     @Override
-    public void writeToParcel(Parcel parcel, int i)
-    {
+    public void writeToParcel(Parcel parcel, int i) {
         parcel.writeString(id);
         parcel.writeString(author);
         parcel.writeString(content);
