@@ -7,7 +7,6 @@ import dagger.Subcomponent;
  */
 @DetailsScope
 @Subcomponent(modules = {DetailsModule.class})
-public interface DetailsComponent
-{
+public interface DetailsComponent {
     void inject(MovieDetailsFragment target);
 }
