@@ -1,30 +1,26 @@
 package com.esoxjem.movieguide.details;
 
 import android.os.Bundle;
-import android.support.v4.app.NavUtils;
-import android.support.v7.app.AppCompatActivity;
 import android.view.MenuItem;
 
-import com.esoxjem.movieguide.R;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.core.app.NavUtils;
+
 import com.esoxjem.movieguide.Constants;
 import com.esoxjem.movieguide.Movie;
+import com.esoxjem.movieguide.R;
 
-public class MovieDetailsActivity extends AppCompatActivity
-{
+public class MovieDetailsActivity extends AppCompatActivity {
     @Override
-    protected void onCreate(Bundle savedInstanceState)
-    {
+    protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_movie_details);
 
-        if (savedInstanceState == null)
-        {
+        if (savedInstanceState == null) {
             Bundle extras = getIntent().getExtras();
-            if (extras != null && extras.containsKey(Constants.MOVIE))
-            {
+            if (extras != null && extras.containsKey(Constants.MOVIE)) {
                 Movie movie = extras.getParcelable(Constants.MOVIE);
-                if (movie != null)
-                {
+                if (movie != null) {
                     MovieDetailsFragment movieDetailsFragment = MovieDetailsFragment.getInstance(movie);
                     getSupportFragmentManager().beginTransaction().add(R.id.movie_details_container, movieDetailsFragment).commit();
                 }
@@ -33,10 +29,8 @@ public class MovieDetailsActivity extends AppCompatActivity
     }
 
     @Override
-    public boolean onOptionsItemSelected(MenuItem item)
-    {
-        switch (item.getItemId())
-        {
+    public boolean onOptionsItemSelected(MenuItem item) {
+        switch (item.getItemId()) {
             case android.R.id.home:
                 NavUtils.navigateUpFromSameTask(this);
         }
