@@ -10,8 +10,8 @@ import io.reactivex.Observable;
 /**
  * @author arun
  */
-public interface MovieDetailsInteractor
-{
+public interface MovieDetailsInteractor {
     Observable<List<Video>> getTrailers(String id);
+
     Observable<List<Review>> getReviews(String id);
 }
