@@ -5,8 +5,7 @@ import com.esoxjem.movieguide.Movie;
 /**
  * @author arun
  */
-public interface MovieDetailsPresenter
-{
+public interface MovieDetailsPresenter {
     void showDetails(Movie movie);
 
     void showTrailers(Movie movie);
