@@ -7,10 +7,12 @@ import java.util.List;
 /**
  * @author arun
  */
-public interface FavoritesInteractor
-{
+public interface FavoritesInteractor {
     void setFavorite(Movie movie);
+
     boolean isFavorite(String id);
+
     List<Movie> getFavorites();
+
     void unFavorite(String id);
 }
