@@ -2,49 +2,35 @@ package com.esoxjem.movieguide.favorites;
 
 import com.esoxjem.movieguide.Movie;
 
-import java.io.IOException;
-import java.util.ArrayList;
 import java.util.List;
 
 /**
  * @author arun
  */
-class FavoritesInteractorImpl implements FavoritesInteractor
-{
+class FavoritesInteractorImpl implements FavoritesInteractor {
     private FavoritesStore favoritesStore;
 
-    FavoritesInteractorImpl(FavoritesStore store)
-    {
+    FavoritesInteractorImpl(FavoritesStore store) {
         favoritesStore = store;
     }
 
     @Override
-    public void setFavorite(Movie movie)
-    {
+    public void setFavorite(Movie movie) {
         favoritesStore.setFavorite(movie);
     }
 
     @Override
-    public boolean isFavorite(String id)
-    {
+    public boolean isFavorite(String id) {
         return favoritesStore.isFavorite(id);
     }
 
     @Override
-    public List<Movie> getFavorites()
-    {
-        try
-        {
-            return favoritesStore.getFavorites();
-        } catch (IOException ignored)
-        {
-            return new ArrayList<>(0);
-        }
+    public List<Movie> getFavorites() {
+        return favoritesStore.getFavorites();
     }
 
     @Override
-    public void unFavorite(String id)
-    {
+    public void unFavorite(String id) {
         favoritesStore.unfavorite(id);
     }
 }
