@@ -11,12 +11,10 @@ import dagger.Provides;
  * @author pulkitkumar
  */
 @Module(includes = AppModule.class)
-public class FavoritesModule
-{
+public class FavoritesModule {
     @Provides
     @Singleton
-    FavoritesInteractor provideFavouritesInteractor(FavoritesStore store)
-    {
+    FavoritesInteractor provideFavouritesInteractor(FavoritesStore store) {
         return new FavoritesInteractorImpl(store);
     }
 }
