@@ -10,8 +10,7 @@ import dagger.Subcomponent;
  */
 @ListingScope
 @Subcomponent(modules = {ListingModule.class, SortingModule.class})
-public interface ListingComponent
-{
+public interface ListingComponent {
     MoviesListingFragment inject(MoviesListingFragment fragment);
 
     SortingDialogFragment inject(SortingDialogFragment fragment);
