@@ -9,8 +9,10 @@ import io.reactivex.Observable;
 /**
  * @author arun
  */
-public interface MoviesListingInteractor
-{
+public interface MoviesListingInteractor {
     boolean isPaginationSupported();
+
     Observable<List<Movie>> fetchMovies(int page);
+
+    Observable<List<Movie>> searchMovie(String searchQuery);
 }
