@@ -3,13 +3,16 @@ package com.esoxjem.movieguide.listing;
 /**
  * @author arun
  */
-public interface MoviesListingPresenter
-{
+public interface MoviesListingPresenter {
     void firstPage();
 
     void nextPage();
 
     void setView(MoviesListingView view);
+
+    void searchMovie(String searchText);
+
+    void searchMovieBackPressed();
 
     void destroy();
 }
