@@ -7,10 +7,12 @@ import java.util.List;
 /**
  * @author arun
  */
-interface MoviesListingView
-{
+interface MoviesListingView {
     void showMovies(List<Movie> movies);
+
     void loadingStarted();
+
     void loadingFailed(String errorMessage);
+
     void onMovieClicked(Movie movie);
 }
