@@ -4,13 +4,14 @@ import android.annotation.SuppressLint;
 import android.app.Dialog;
 import android.content.Context;
 import android.os.Bundle;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
-import android.support.v4.app.DialogFragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.widget.RadioButton;
 import android.widget.RadioGroup;
+
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.fragment.app.DialogFragment;
 
 import com.esoxjem.movieguide.BaseApplication;
 import com.esoxjem.movieguide.R;
@@ -26,8 +27,7 @@ import butterknife.Unbinder;
 /**
  * @author arun
  */
-public class SortingDialogFragment extends DialogFragment implements SortingDialogView, RadioGroup.OnCheckedChangeListener
-{
+public class SortingDialogFragment extends DialogFragment implements SortingDialogView, RadioGroup.OnCheckedChangeListener {
     @Inject
     SortingDialogPresenter sortingDialogPresenter;
 
@@ -45,15 +45,13 @@ public class SortingDialogFragment extends DialogFragment implements SortingDial
     private static MoviesListingPresenter moviesListingPresenter;
     private Unbinder unbinder;
 
-    public static SortingDialogFragment newInstance(MoviesListingPresenter moviesListingPresenter)
-    {
+    public static SortingDialogFragment newInstance(MoviesListingPresenter moviesListingPresenter) {
         SortingDialogFragment.moviesListingPresenter = moviesListingPresenter;
         return new SortingDialogFragment();
     }
 
     @Override
-    public void onCreate(@Nullable Bundle savedInstanceState)
-    {
+    public void onCreate(@Nullable Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setRetainInstance(true);
         ((BaseApplication) getActivity().getApplication()).getListingComponent().inject(this);
@@ -62,8 +60,7 @@ public class SortingDialogFragment extends DialogFragment implements SortingDial
 
     @NonNull
     @Override
-    public Dialog onCreateDialog(Bundle savedInstanceState)
-    {
+    public Dialog onCreateDialog(Bundle savedInstanceState) {
         LayoutInflater inflater = (LayoutInflater) getActivity().getSystemService(Context.LAYOUT_INFLATER_SERVICE);
         View dialogView = inflater.inflate(R.layout.sorting_options, null);
         unbinder = ButterKnife.bind(this, dialogView);
@@ -76,43 +73,36 @@ public class SortingDialogFragment extends DialogFragment implements SortingDial
         return dialog;
     }
 
-    private void initViews()
-    {
+    private void initViews() {
         sortingDialogPresenter.setLastSavedOption();
         sortingOptionsGroup.setOnCheckedChangeListener(this);
     }
 
     @Override
-    public void setPopularChecked()
-    {
+    public void setPopularChecked() {
         mostPopular.setChecked(true);
     }
 
 
     @Override
-    public void setNewestChecked()
-    {
+    public void setNewestChecked() {
         newest.setChecked(true);
     }
 
     @Override
-    public void setHighestRatedChecked()
-    {
+    public void setHighestRatedChecked() {
         highestRated.setChecked(true);
     }
 
     @Override
-    public void setFavoritesChecked()
-    {
+    public void setFavoritesChecked() {
         favorites.setChecked(true);
     }
 
     @SuppressLint("InvalidR2Usage")
     @Override
-    public void onCheckedChanged(RadioGroup radioGroup, int checkedId)
-    {
-        switch (checkedId)
-        {
+    public void onCheckedChanged(RadioGroup radioGroup, int checkedId) {
+        switch (checkedId) {
             case R2.id.most_popular:
                 sortingDialogPresenter.onPopularMoviesSelected();
                 moviesListingPresenter.firstPage();
@@ -135,14 +125,12 @@ public class SortingDialogFragment extends DialogFragment implements SortingDial
     }
 
     @Override
-    public void dismissDialog()
-    {
+    public void dismissDialog() {
         dismiss();
     }
 
     @Override
-    public void onDestroyView()
-    {
+    public void onDestroyView() {
         super.onDestroyView();
         sortingDialogPresenter.destroy();
         unbinder.unbind();
