@@ -3,8 +3,7 @@ package com.esoxjem.movieguide.listing.sorting;
 /**
  * @author arun
  */
-public interface SortingDialogInteractor
-{
+public interface SortingDialogInteractor {
     int getSelectedSortingOption();
 
     void setSortingOption(SortType sortType);
