@@ -3,8 +3,7 @@ package com.esoxjem.movieguide.listing.sorting;
 /**
  * @author arun
  */
-public interface SortingDialogPresenter
-{
+public interface SortingDialogPresenter {
     void setLastSavedOption();
 
     void onPopularMoviesSelected();
