@@ -4,9 +4,9 @@ import com.esoxjem.movieguide.MoviesWraper;
 import com.esoxjem.movieguide.ReviewsWrapper;
 import com.esoxjem.movieguide.VideoWrapper;
 
+import io.reactivex.Observable;
 import retrofit2.http.GET;
 import retrofit2.http.Path;
-import io.reactivex.Observable;
 import retrofit2.http.Query;
 
 /**
@@ -22,12 +22,15 @@ public interface TmdbWebService {
     Observable<MoviesWraper> highestRatedMovies(@Query("page") int page);
 
     @GET("3/discover/movie?language=en&sort_by=release_date.desc")
-    Observable<MoviesWraper> newestMovies(@Query("release_date.lte") String maxReleaseDate,@Query("vote_count.gte") int minVoteCount);
+    Observable<MoviesWraper> newestMovies(@Query("release_date.lte") String maxReleaseDate, @Query("vote_count.gte") int minVoteCount);
 
     @GET("3/movie/{movieId}/videos")
     Observable<VideoWrapper> trailers(@Path("movieId") String movieId);
 
     @GET("3/movie/{movieId}/reviews")
     Observable<ReviewsWrapper> reviews(@Path("movieId") String movieId);
+
+    @GET("3/search/movie?language=en-US&page=1")
+    Observable<MoviesWraper> searchMovies(@Query("query") String searchQuery);
 
 }
