@@ -6,20 +6,15 @@ import io.reactivex.disposables.Disposable;
 /**
  * @author arun
  */
-public class RxUtils
-{
-    public static void unsubscribe(Disposable subscription)
-    {
-        if (subscription != null && !subscription.isDisposed())
-        {
+public class RxUtils {
+    public static void unsubscribe(Disposable subscription) {
+        if (subscription != null && !subscription.isDisposed()) {
             subscription.dispose();
         } // else subscription doesn't exist or already unsubscribed
     }
 
-    public static void unsubscribe(Disposable... subscriptions)
-    {
-        for (Disposable subscription : subscriptions)
-        {
+    public static void unsubscribe(Disposable... subscriptions) {
+        for (Disposable subscription : subscriptions) {
             unsubscribe(subscription);
         }
     }
